+++ src/components/index.js
@@ -9,4 +9,4 @@ export { default as Footer } from './footer/Footer';
 export { default as DownloadButtons } from './download_buttons/DownloadButtons.jsx';
 export { default as Card } from './card/Card';
 export { default as Modal } from './modal/Modal';
-export { default as NewsletterSignup } from './newsletter-signup/NewsletterSignup';
+export { default as NewsletterSignupModal } from './newsletter-signup-modal/NewsletterSignupModal';
